refactor(tab-manager): organize components into chat/ and tabs/ subdirectories
