+++ apps/tab-manager/src/lib/components/chat/AiChat.svelte
