+++ apps/tab-manager/src/lib/components/chat/ChatErrorBanner.svelte
