+++ apps/tab-manager/src/lib/components/chat/ChatInput.svelte
