+++ apps/tab-manager/src/lib/components/chat/ConversationPicker.svelte
