+++ apps/tab-manager/src/lib/components/chat/MessageParts.svelte
