+++ apps/tab-manager/src/lib/components/chat/ProviderSelect.svelte
