+++ apps/tab-manager/src/lib/components/chat/ToolResultPart.svelte
