+++ apps/tab-manager/src/lib/components/tabs/FlatTabList.svelte
