+++ apps/tab-manager/src/lib/components/tabs/TabFavicon.svelte
