+++ apps/tab-manager/src/lib/components/tabs/TabItem.svelte
